+++ trunk/linbox/linbox/algorithms/cia.h
@@ -60,7 +60,7 @@ namespace LinBox
 		RandomPrimeIterator primeg (22);
 		++primeg;
 		Field F(*primeg);
-		FBlackbox fbb(F, A.rowdim(), A.coldim());
+		FBlackbox fbb(F, (int)A.rowdim(), (int)A.coldim());
 		MatrixHom::map(fbb, A, F);
 		charpoly (fieldCharPoly, fbb, M);
 		/* Determination of the multiplicities */
